BusTripStopモデルを別ファイルに切り出し
